# docs/modbus_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API文档",
        "title": "物联大师网关接口文档",
        "contact": {},
        "version": "1.0 版本"
    },
    "basePath": "/api/gateway/api/",
    "paths": {
...
        "/server/count": {
            "post": {
                "description": "查询服务器数量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "查询服务器数量",
                "parameters": [
                    {
                        "description": "查询参数",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ParamSearch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-int64"
                        }
                    }
                }
            }
        },
        "/server/create": {
            "post": {
                "description": "创建服务器",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "创建服务器",
                "parameters": [
                    {
                        "description": "服务器信息",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Server"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Server"
                        }
                    }
                }
            }
        },
...
        "/server/import": {
            "post": {
                "description": "导入服务端",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "导入服务端",
                "parameters": [
                    {
                        "type": "file",
                        "description": "压缩包",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-int64"
                        }
                    }
                }
            }
        },
        "/server/list": {
            "get": {
                "description": "查询服务器",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "查询服务器",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "skip",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyList-types_Server"
                        }
                    }
                }
            }
        },
        "/server/search": {
            "post": {
                "description": "查询服务器",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "查询服务器",
                "parameters": [
                    {
                        "description": "查询参数",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ParamSearch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyList-types_Server"
                        }
                    }
                }
            }
        },
        "/server/{id}": {
            "get": {
                "description": "获取服务器",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "获取服务器",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "服务器ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Server"
                        }
                    }
                }
            },
            "post": {
                "description": "修改服务器",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "修改服务器",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "服务器ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "服务器信息",
                        "name": "server",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Server"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Server"
                        }
                    }
                }
            }
        },
        "/server/{id}/delete": {
            "get": {
                "description": "删除服务器",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "删除服务器",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "服务器ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Server"
                        }
                    }
                }
            }
        },
        "/server/{id}/disable": {
            "get": {
                "description": "禁用服务器",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "禁用服务器",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "服务器ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Server"
                        }
                    }
                }
            }
        },
        "/server/{id}/enable": {
            "get": {
                "description": "启用服务器",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "启用服务器",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "服务器ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Server"
                        }
                    }
                }
            }
        },
        "/server/{id}/start": {
            "get": {
                "description": "启动服务端",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "启动服务端",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "服务端ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Server"
                        }
                    }
                }
            }
        },
        "/server/{id}/stop": {
            "get": {
                "description": "停止服务端",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "停止服务端",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "服务端ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Server"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ParamSearch": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "object",
                    "additionalProperties": true
                },
                "keyword": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "skip": {
                    "type": "integer"
                },
                "sort": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        },
        "api.ReplyData-int64": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "api.ReplyData-types_Server": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.Server"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "api.ReplyList-types_Server": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Server"
                    }
                },
                "error": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "types.DefaultDevice": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "product_id": {
                    "description": "产品ID",
                    "type": "string"
                },
                "slave": {
                    "description": "从站号",
                    "type": "integer"
                }
            }
        },
...
        "types.Server": {
            "type": "object",
            "properties": {
                "created": {
                    "description": "创建时间",
                    "type": "string"
                },
                "desc": {
                    "description": "说明",
                    "type": "string"
                },
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.DefaultDevice"
                    }
                },
                "disabled": {
                    "type": "boolean"
                },
                "heartbeat": {
                    "description": "心跳包",
                    "type": "string"
                },
                "id": {
                    "description": "ID",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "poller_interval": {
                    "description": "采集间隔",
                    "type": "integer"
                },
                "poller_period": {
                    "description": "采集周期",
                    "type": "integer"
                },
                "port": {
                    "description": "监听端口",
                    "type": "integer"
                },
                "protocol_name": {
                    "description": "协议 rtu tcp parallel-tcp",
                    "type": "string"
                },
                "protocol_options": {
                    "description": "协议参数",
                    "type": "string"
                },
                "running": {
                    "type": "boolean"
                },
                "standalone": {
                    "description": "单例模式（不支持注册）",
                    "type": "boolean"
                }
            }
        }
    }
}